start working on 1.2.0
so i've had most of this locally saved because i didnt know how git branches worked. this commit migrates to a proper venv (poetry), stops downloading the icon and template config instead choosing to create them locally, and switches the config format from json to toml. also beginning of the tagging feature, and storing nation list as a tuple instead of a list.